Use universal dependencies tagset by default.

+++ src/tag/PhemTag.cpp
@@ -1,14 +1,14 @@
 #include "PhemTag.h"
 
 namespace base {
-static const boost::bimap<uint128_t, std::string> PHEM_MAP =
-    boost::assign::list_of<boost::bimap<uint128_t, std::string>::relation>(0x00, "UNKN")(0x01, "PREFIX")(0x02, "ROOT")(0x04, "SUFFIX")(0x08, "ENDING");
+static const boost::bimap<uint64_t, std::string> PHEM_MAP =
+    boost::assign::list_of<boost::bimap<uint64_t, std::string>::relation>(0x00, "UNKN")(0x01, "PREFIX")(0x02, "ROOT")(0x04, "SUFFIX")(0x08, "ENDING");
 
-const PhemTag PhemTag::UNKN(uint128_t(0x00));
-const PhemTag PhemTag::PREFIX(uint128_t(0x01));
-const PhemTag PhemTag::ROOT(uint128_t(0x02));
-const PhemTag PhemTag::SUFFIX(uint128_t(0x04));
-const PhemTag PhemTag::ENDING(uint128_t(0x08));
+const PhemTag PhemTag::UNKN(uint64_t(0x00));
+const PhemTag PhemTag::PREFIX(uint64_t(0x01));
+const PhemTag PhemTag::ROOT(uint64_t(0x02));
+const PhemTag PhemTag::SUFFIX(uint64_t(0x04));
+const PhemTag PhemTag::ENDING(uint64_t(0x08));
 
 const std::vector<PhemTag> PhemTag::inner_runner = {
     UNKN,
@@ -18,7 +18,7 @@ const std::vector<PhemTag> PhemTag::inner_runner = {
     ENDING,
 };
 
-PhemTag::PhemTag(uint128_t val)
+PhemTag::PhemTag(uint64_t val)
     : ITag(val, &PHEM_MAP) {
 }
 
@@ -27,6 +27,6 @@ PhemTag::PhemTag(const std::string& val)
 }
 
 PhemTag::PhemTag()
-    : ITag((uint128_t)0, &PHEM_MAP) {
+    : ITag((uint64_t)0, &PHEM_MAP) {
 }
 }
